update listner with abstarct class
